+++ changelog.md
@@ -23,6 +23,7 @@ Codion Change Log
 - AttributeDefinition.Builder.valueRange(), minimumValue() and maximumValue() renamed range(), minimum() and maximum().
 - AttributeDefinition.maximumFractionDigits() and decimalRoundingMode() renamed to fractionDigits() and roundingMode().
 - AttributeDefinition.Builder.numberFormatGrouping() renamed numberGrouping().
+- AttributeDefinition.Builder.localeDateTimePattern() renamed dateTimePattern().
 ### is.codion.plugin.flatlaf.intellij
 - Catppuccin themes added.
 - VSCode themes added.